+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next Release
 
+- Fix: changing a character's sheet now properly updates their move sheet
 - Tours: the "welcome" tour now skips the 5th step if there's no scene, avoiding a dangling ending.
 - Tours: add a tour for the Ironsworn character sheet ([#657](https://github.com/ben/foundry-ironsworn/pull/657))